Move CreatePublicationRequest to publication-rest
